doc: explain how to use paperToaster service
I'm not sure this is the best way to explain this, but I felt that the previous documentation did not offer any clues as to *how* to use the `{{paper-toaster}}` component. Someone kindly helped me out via Discord chat saying that `{{paper-toaster}}` needs to be added to the application template.